Clustering trust_domain is broken due to api changes in tm.sys.application
Issues:
Fixes #521

Problem:
The clustering module trust_domain.py is broken because the api changed
for tm.sys.applications to tm.sys.application. I will fix the issue and
submit the fix, and we'll have to spin another release to remedy this.

Analysis:
Modified the references to applications in trust_domain.py and reran the
clustering tests.

Tests:
Clustering tests passed locally.